outlier reduction from ROOT1

+++ train_gpt.py
@@ -656,7 +656,7 @@ class NorMuon(torch.optim.Optimizer):
             if num_params == 0:
                 v_chunk = updated_grads
             else:
-                threshold = updated_grads.view(updated_grads.shape[0],-1).abs().quantile(0.9)
+                threshold = updated_grads.view(updated_grads.shape[0],-1).abs().float().quantile(0.9)
                 updated_grads.sub_(F.softshrink(updated_grads, threshold))
                 v_chunk = polar_express(updated_grads, split_baddbmm=(ref_param.label == 'mlp'))
